feat : get id curr user for user top genre book

+++ ENSIKLO/ENSIKLO/ENSIKLO/Services/APIBookService.cs
@@ -76,7 +76,7 @@ namespace ENSIKLO.Services
             return JsonSerializer.Deserialize<IEnumerable<Book>>(responseAsString);
         }
 
-        public async Task<IEnumerable<Book>> GetUserTopGenreBook(int id_user, int limit, bool forceRefresh = false)
+        public async Task<IEnumerable<Book>> GetUserTopGenreBook(Int64 id_user, int limit, bool forceRefresh = false)
         {
             var response = await _httpClient.GetAsync($"Book/TopGenre/{id_user}/{limit}");
